resolved all errors coming due to eventDetails api changes, now all events details data coming successfully

+++ app/src/main/java/com/softaai/footballmatchbooking/data/model/eventdetails/EventDetail.kt
@@ -27,7 +27,7 @@ data class EventDetail(
     @Json(name = "name")
     val name: String,
     @Json(name = "price")
-    val price: Int,
+    val price: Float,
     @Json(name = "sport")
     val sport: String,
     @Json(name = "tags")
@@ -37,7 +37,7 @@ data class EventDetail(
     @Json(name = "ticketsSold")
     val ticketsSold: Int,
     @Json(name = "totalPrize")
-    val totalPrize: Int,
+    val totalPrize: Float,
     @Json(name = "venueInformation")
     val venueInformation: String
 )
